requirements.txt
Django==1.6.2
South==0.8.4
Whoosh==2.6.0
argparse==1.2.1
django-appconf==0.6
django-autocomplete-light==1.4.13
django-autoslug==1.7.2
django-braces==1.3.1
django-compressor==1.3
django-floppyforms==1.1.1
django-gravatar2==1.1.3
django-haystack==2.1.0
django-model-utils==2.0.2
django-pipeline==1.3.21
django-rest-swagger==0.1.13
djangorestframework==2.3.12
djangorestframework-csv==1.3.0
futures==2.1.6
gunicorn==18.0
python-dateutil==2.2
six==1.5.2
wsgiref==0.1.2
